+++ src/dapp/CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The output from `dapp build` now uses color to differentiate warnings and errors
 - Support for solc 0.8.9/0.8.8
 
+- `dapp snapshot` and `dapp check-snapshot` commands to snapshot test gas usage.
+
 ### Changed
 
 - Dapp remappings ignores non-directories in `DAPP_LIB`